test_ns1/requirements.txt
pytest
requests
paho-mqtt
